Refactoring CreateQueryIndexOperation.swift to move each class/struct to it's own file
